Ejector working... mostly
Ejector seems to be working correctly, except... When I fire the gun, it ejects a shell as expected. However, when I manually pull the bolt back and eject a non-fired round, it ejects it in a random direction with extreme force. No idea why, layers are set so it doesn't interact with the collider and the rigidbody and collider on both cases are exactly identical.