Bullet drop zones fixed!
This has been a bit of a roadblock for me, individual rounds weren't snapping into dropzones correctly. Decided to change the way I went about it, utilising events instead of inheriting scripts. Works now!

Need to get magazines set up for rounds visibly shown in the mag.  I'm going to have a user-set amount of shown rounds, so that not every round has to move position with every shot. That sounds a bit confusing... It'll make sense when I implement it, haha.