Updated to 2017.1, built a pistol, disable model on grab
Updated my copy of Unity to 2017.1. Doesn't seem to cause any issues, but if anyone encounters any issues let me know.

Built an example pistol, it's in the example scene.

Disabling controller model on grab is now possible, but does not work unless you assign something in VRTK's InteractGrab script's Controller Attach Point (usually found on Controller Aliases). I suggest adding a kinematic rigidbody to each controller's object, and assigning that as the controller attach point. The issue is, if it's not explicitly assigned, the attach point is actually a part of the controller model, so disabling the model also disables the weapon.