Bullet drop zones in for generic IS
That's the last piece that needs to be done before release! Now I'm going to build an example scene for both VRTK and non-VRTK, get the readme and changelog fixed up, and push!

The store takes a few days to approve it, so I'm gonna spend that time making new promo material and finish up documentation.

If anyone finds bugs - anything, big or small - please let me know!