Centralized events to Weapon.cs
For ease of use with other scripts, events have been centralized to Weapon.cs, instead of having to search for the component that fires an event.

Also some editor tweaks - I found myself forgetting to set up the bolt controller pretty frequently, so I set it to not disable the mesh renderer, and have a red material assigned to it. Also added a button that disables the mesh renderer once finished setting it up.