Magazine fixes, realistic projectile-based bullet
Fixed an issue with complex magazines not pushing rounds correctly, and bolts being able to fire while the chamber was open.

Added a script and interface for realistic bullet physics. RealisticBulletProjectile.cs is intended to be placed on the actual projectile object, and allows the bullet to send an attack to anything it hits (using linecasting, so high-speed projectiles still work since Unity's collision detection has issues with high speeds)